Move "@types/luxon" to the "dependencies" section
The exiftool-vendored uses the luxon's DateTime type in public interfaces (for example, in ExifTime class). So, the @types/luxon should be in the "dependencies" section instead of "devDependencies". Otherwise, the TS code that uses exiftool-vendored and doesn't use luxon in its own dependencies, will not have access to the luxon's types.